requirements.txt
httpx>=0.27
playwright>=1.44
beautifulsoup4>=4.12
sqlalchemy>=2.0
alembic>=1.13
dash>=2.17
dash-bootstrap-components>=1.6
plotly>=5.22
pandas>=2.2
numpy>=1.26
statsmodels>=0.14
apscheduler>=3.10
pyyaml>=6.0
python-dotenv>=1.0
haversine>=2.8
rich>=13.7
click>=8.1
pytest>=8.2
